// frontend/src/api/model/todoCreate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * FastAPI React Backend
 * OpenAPI spec version: 0.1.0
 */
import type { TodoCreateDescription } from './todoCreateDescription';

/**
 * Todoアイテム作成用のモデル
 */
export interface TodoCreate {
  title: string;
  description?: TodoCreateDescription;
  completed?: boolean;
}