// frontend/src/api/model/todoResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * FastAPI React Backend
 * OpenAPI spec version: 0.1.0
 */
import type { TodoResponseDescription } from './todoResponseDescription';
import type { TodoResponseUpdatedAt } from './todoResponseUpdatedAt';

/**
 * Todoアイテムのレスポンスモデル
 */
export interface TodoResponse {
  id: string;
  title: string;
  description?: TodoResponseDescription;
  completed: boolean;
  created_at: string;
  updated_at?: TodoResponseUpdatedAt;
}