// frontend/src/api/model/todoUpdate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * FastAPI React Backend
 * OpenAPI spec version: 0.1.0
 */
import type { TodoUpdateTitle } from './todoUpdateTitle';
import type { TodoUpdateDescription } from './todoUpdateDescription';
import type { TodoUpdateCompleted } from './todoUpdateCompleted';

/**
 * Todoアイテム更新用のモデル
 */
export interface TodoUpdate {
  title?: TodoUpdateTitle;
  description?: TodoUpdateDescription;
  completed?: TodoUpdateCompleted;
}